Added HsCmd data type, started refactoring
I added the HsCmd data type to HsExpr.  This means that I've also
updated the match and stmt data types to be parametrized so they
can either have bodies of exprs or cmds.

I also started updating the rest of the compiler to use these new
types.  Currently, HsExpr type-checks, but there are errors in
HsUtils (and potentially other files further in the build order).