Addressing problems in Parser
I added HsArrForm and HsArrApp back into HsExpr because of the way
the parser works.  Currently, commands are still parsed as exprs,
but then they are immediately changed in RdrHsSyn.  That said, we
still need the extra forms for HsExpr for the initial parsing.

This commit also includes some fixes and updates with getting
the HsCmd types working properly.